+++ ephyr_control/instance/instance.py
@@ -5,9 +5,10 @@ __all__ = ("EphyrInstance",)
 from typing import Optional
 
 from ephyr_control.instance.protocols import EphyrInstanceProtocol
+from ephyr_control.types import EphyrConfig
 
 
-@dataclasses.dataclass
+@dataclasses.dataclass(unsafe_hash=True)
 class EphyrInstance(EphyrInstanceProtocol):
     """Ephyr instance - represents one server with its address.
     Connects to server using provided arguments and can perform simple actions on it.
@@ -26,6 +27,10 @@ class EphyrInstance(EphyrInstanceProtocol):
     password: Optional[str] = None
     https: bool = True
 
+    @classmethod
+    def from_config(cls, config: EphyrConfig) -> "EphyrInstance":
+        return cls(**config)
+
     @property
     def ip(self) -> str:
         return self.ipv4
@@ -40,12 +45,16 @@ class EphyrInstance(EphyrInstanceProtocol):
 
     @property
     def scheme(self) -> str:
-        return "https" if self.https else "http"
+        return "https" if self.https and self.domain else "http"
+
+    def address(self, with_password=False):
+        if with_password:
+            return f"{self.scheme}://1:{self.password}@{self.host}/"
+        return f"{self.scheme}://{self.host}/"
 
     def print(self):
         """Print data about itself."""
-        host = self.domain if self.domain else self.ipv4
-        address = f"{self.scheme}://{host}/"
         extra = " (" + str(self.ipv4) + ")" if self.domain else ""
-        print(f'Ephyr instance "{self.title}" at {address}{extra}')
+        print(f'Ephyr instance "{self.title}" at {self.address()}{extra}')
+        print(f"Ephyr instance URL with password: {self.address(with_password=True)}")
         print(f"Ephyr suggested password: {self.password}")