+++ ephyr_control/state/restream/input/input_source.py
@@ -1,7 +1,12 @@
 import dataclasses
 from typing import List
 
-from .failover_input import FailoverInput, backup_input_factory, main_input_factory
+from .failover_input import (
+    FailoverInput,
+    backup1_input_factory,
+    backup2_input_factory,
+    primary_input_factory,
+)
 
 __all__ = ("InputSource",)
 
@@ -9,10 +14,18 @@ __all__ = ("InputSource",)
 @dataclasses.dataclass
 class InputSource:
     failover_inputs: List[FailoverInput] = dataclasses.field(
-        default_factory=lambda: [main_input_factory(), backup_input_factory()]
+        default_factory=lambda: [
+            primary_input_factory(),
+            backup1_input_factory(),
+            backup2_input_factory(),
+        ]
     )
 
-    FI_KEYS_DEFAULT = (FailoverInput.KEY_MAIN, FailoverInput.KEY_BACKUP)
+    FI_KEYS_DEFAULT = (
+        (FailoverInput.KEY_PRIMARY, None),
+        (FailoverInput.KEY_BACKUP1, None),
+        (FailoverInput.KEY_BACKUP2, None),
+    )
 
     def __post_init__(self):
         # ensure unique restream keys
@@ -28,17 +41,21 @@ class InputSource:
             raise KeyError(f'FailoverInput with key="{foinput_key}" not found.')
 
     @property
-    def main_input(self) -> FailoverInput:
+    def primary_input(self) -> FailoverInput:
         return self.failover_inputs[0]
 
     @property
-    def backup_input(self) -> FailoverInput:
+    def backup1_input(self) -> FailoverInput:
         return self.failover_inputs[1]
+
+    @property
+    def backup2_input(self) -> FailoverInput:
+        return self.failover_inputs[2]
 
     @classmethod
     def with_random_keys(
         cls,
-        key_prefixes: [str, str] = FI_KEYS_DEFAULT,
+        key_prefixes_with_labels: List[List[str]] = FI_KEYS_DEFAULT,
         key_random_chars: int = FailoverInput.KEY_RANDOM_LENGTH_DEFAULT,
     ) -> "InputSource":
         return cls(
@@ -46,7 +63,8 @@ class InputSource:
                 FailoverInput.with_random_key(
                     key_prefix=key_prefix,
                     key_random_chars=key_random_chars,
+                    endpoint_label=endpoint_label,
                 )
-                for key_prefix in key_prefixes
+                for (key_prefix, endpoint_label) in key_prefixes_with_labels
             ]
         )